Clean up documentation structure:
Remove bloat by replacing the system of nested directories with
`README.md` in each in the syntax documentation to a flat structure of
multiple .md files in the `syntax/` directory.

Move style and implementation related documentation outside of the
`syntax/` directory.

Correct internal links at the top of each .md file.